+++ client/components/Dropdown/DropdownMenu.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React, { forwardRef, useCallback, useRef, useState } from 'react';
 import useModalClose from '../../common/useModalClose';
 import DownArrowIcon from '../../images/down-filled-triangle.svg';
-import { DropdownWrapper } from '../Dropdown';
+import { DropdownWrapper } from './DropdownWrapper';
 
 // TODO: enable arrow keys to navigate options from list
